services/crawler/pyproject.toml
[project]
name = "hokiehelp-crawler"
version = "0.1.0"
description = "Web crawler for cs.vt.edu that stores Markdown in MinIO"
requires-python = ">=3.11"
dependencies = [
    "crawl4ai>=0.6.0",
    "minio>=7.2.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
hokiehelp-crawl = "crawler.main:cli"
